Initial Rodroid commit
This patch adds an Android client for RoDI (Robot Didáctico Inalámbrico)
to remote control the robot, by using its default firmware web services.

For now only the move command is implemented, but the plan is to support
as followup all the operations that are offered by the web services.